% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_settings_importer.R
\name{read_settings_from_fileInput}
\alias{read_settings_from_fileInput}
\title{Read settings from json files}
\usage{
read_settings_from_fileInput(file_df)
}
\arguments{
\item{file_df}{the file dataframe returned by import_settings fileInput}
}
\value{
a list of settings
}
\description{
Read settings from json files
}
